DCS-969 remove dependency on bunyan-request-logger
This was pulling in an old version of snyk as a transient dependency which was reporting a vulnerability.

The actual code we were relying on was really trivial so just using that instead
