--- backend/setupCurrentUserAndRequestLogging.js
@@ -1,22 +1,0 @@
-const express = require('express')
-const bunyanRequestLogger = require('bunyan-request-logger')
-
-const loggingSerialiser = require('./loggingSerialiser')
-const currentUser = require('./middleware/currentUser')
-
-const router = express.Router()
-
-module.exports = ({ oauthApi }) => {
-  router.use(currentUser({ oauthApi }))
-  router.use(async (req, res, next) => {
-    res.locals = {
-      ...res.locals,
-      currentUrlPath: req.originalUrl,
-    }
-    next()
-  })
-
-  router.use(bunyanRequestLogger({ name: 'Book video link http', serializers: loggingSerialiser }).requestLogger())
-
-  return router
-}
