+++ package.json
@@ -16,7 +16,6 @@
     "axios": "0.21.1",
     "bunyan": "^1.8.15",
     "bunyan-format": "^0.2.1",
-    "bunyan-request-logger": "^2.1.0",
     "connect-flash": "^0.1.1",
     "connect-redis": "^5.1.0",
     "cookie-parser": "^1.4.5",
@@ -47,6 +46,7 @@
     "universal-analytics": "^0.4.23"
   },
   "devDependencies": {
+    "@types/bunyan": "^1.8.6",
     "@types/cookie-parser": "^1.4.2",
     "@types/express": "^4.17.11",
     "@types/express-session": "^1.17.3",
